requirements.txt
langchain==0.3.4
openai==1.52.2
openpyxl==3.1.5
pandas==2.2.3
PyPDF2==3.0.1
python-dotenv==1.0.1
streamlit==1.38.0
